create SoundFiltersConfig
